Updated button and refactored storybook
