+++ src/Button/Button.tsx
@@ -4,12 +4,18 @@ import '../main.scss';
 
 const Button = ({
   label,
+  onClick = () => {},
   className = '',
 }: {
   label: String;
+  onClick: () => void;
   className?: string;
 }) => {
-  return <button className={`lilypad-button ${className}`}>{label}</button>;
+  return (
+    <button className={`lilypad-button ${className}`} onClick={onClick}>
+      {label}
+    </button>
+  );
 };
 
 export default Button;
