--- src/Row/stories/story.scss
@@ -1,7 +1,0 @@
-.lilypad-row {
-  background-color: blue;
-}
-
-.lilypad-flex-container {
-  background-color: red;
-}
