+++ stories/story.scss
@@ -5,3 +5,7 @@
 .lilypad-container {
   background-color: red;
 }
+
+.lilypad-row {
+  background-color: orange;
+}
